refactor(ui): split Composer into picker/menu/layout/enhance modules
- chat/composerAttachments.ts (+tests): clipboard image helpers
- chat/composerLabels.ts (+tests): fmtBytes/fmtInt/clamp01/displayLlmId/
  displayModeLabel
- chat/composerPickerApi.ts: SlashRow/WorkspaceFileRow/PickerFloatRect types
  + slash/at/commands fetches
- chat/AttachedFileChip.tsx: attachment chip + object-URL hook
- chat/useEnhancePrompt.ts: enhance request state + Ctrl+Z restore
- chat/useComposerSheetLayout.ts: sheet/floating-picker measurement (folds
  the duplicated media-query sync effect into one)
- chat/useSlashAtPickers.ts: the full race-guarded / and @ picker state
  machine, choice application and paging
- chat/SlashAtPickerMenus.tsx: picker chrome + portal (slashRowLayoutCss
  contract test repointed here)
- chat/ComposerModeMenus.tsx: mode/model/reasoning portal; owns the model
  filter query (reset per opened menu, as before)
Composer keeps the card render, textarea handlers, attachments state, menu
triggers, context ring, and the #btn-send play SVG (icon contract test
unchanged). Markup, class names, and testids unchanged.
