+++ external/ui/src/ui/chat/Composer.tsx
@@ -8,7 +8,6 @@ import {
   useSyncExternalStore,
 } from "react";
 import type { Dispatch, SetStateAction } from "react";
-import { createPortal } from "react-dom";
 import type { TokenUsage } from "./types";
 import { WorkspaceChips } from "./WorkspaceChips";
 import { useT } from "../i18n/I18nProvider";
@@ -19,208 +18,30 @@ import {
   type ContextBreakdown,
 } from "./ContextBreakdownPopover";
 import { ContextUsageRing } from "./ContextUsageRing";
-import {
-  draftExtendsFailedAtPrefix,
-  atMenuDraftAtCaret,
-} from "../skills/draftAt";
-import {
-  draftExtendsFailedSlashPrefix,
-  slashMenuDraftAtCaret,
-} from "../skills/draftSlash";
-import { filterCommandRows } from "../skills/commandRows";
 import { segmentComposerMirrorSpans } from "../skills/composerMirrorSegments";
-import { workspacePickRowSubtitle } from "../skills/workspacePickRowSubtitle";
 import {
-  pickerRowFromRecent,
-  readWorkspaceAtRecents,
-  recordWorkspaceAtRecent,
-  WORKSPACE_AT_RECENTS_NO_SESSION_KEY,
-} from "../skills/workspaceAtRecents";
-import {
-  shellStackMaxWidthMediaQuery,
   subscribeShellStack,
   snapshotShellStack,
   serverSnapshotShellStack,
 } from "../shellBreakpoint";
 import { contextUsagePercent } from "./contextUsage";
-import {
-  filterLlmModels,
-  groupLlmModelsByVendor,
-  shouldGroupLlmModels,
-  shouldShowLlmFilter,
-} from "./llmModelMenu";
 import { fileTypeIcon } from "../messages/fileTypeIcon";
-
-function fmtBytes(
-  n: number,
-  t: (key: string, params?: Record<string, string | number>) => string,
-): string {
-  if (n < 1024) return t("composer.bytesB", { n });
-  if (n < 1024 * 1024) {
-    return t("composer.bytesKB", { n: (n / 1024).toFixed(1) });
-  }
-  return t("composer.bytesMB", { n: (n / (1024 * 1024)).toFixed(1) });
-}
-
-function clamp01(x: number): number {
-  if (!Number.isFinite(x)) return 0;
-  if (x < 0) return 0;
-  if (x > 1) return 1;
-  return x;
-}
-
-function fmtInt(n: number | undefined): string {
-  if (typeof n !== "number" || !Number.isFinite(n)) return "0";
-  return Math.max(0, Math.trunc(n)).toString();
-}
-
-/** Short label for **`models[].model`** ids (Coddy profile IDs use displayMode elsewhere). */
-function displayLlmId(id: string, fallback: string = "Model"): string {
-  const m = id || "";
-  const i = m.lastIndexOf("/");
-  if (i >= 0 && i < m.length - 1) {
-    return m.slice(i + 1);
-  }
-  return m || fallback;
-}
-
-/** File extension for an image MIME type ("image/svg+xml" -> "svg"). */
-function imageExtFromMime(mime: string): string {
-  const wellKnown: Record<string, string> = {
-    png: "png",
-    jpeg: "jpg",
-    gif: "gif",
-    webp: "webp",
-    "svg+xml": "svg",
-    bmp: "bmp",
-  };
-  const sub = (mime.split("/")[1] || "").toLowerCase();
-  return wellKnown[sub] || sub.replace(/[^a-z0-9]/g, "") || "png";
-}
-
-/** Image files carried by a clipboard **`DataTransfer`** (`kind === "file"` + `image/*` MIME). */
-function clipboardImageFiles(data: DataTransfer | null | undefined): File[] {
-  if (!data || !data.items) return [];
-  const out: File[] = [];
-  for (const item of Array.from(data.items)) {
-    if (item.kind !== "file" || !item.type.startsWith("image/")) continue;
-    const f = item.getAsFile();
-    if (f) out.push(f);
-  }
-  return out;
-}
-
-/**
- * Browsers name every clipboard image "image.png"; give pasted files
- * deterministic per-composer names so chips and session history stay unambiguous.
- */
-function renamePastedImages(
-  files: File[],
-  seqRef: { current: number },
-): File[] {
-  return files.map((f) => {
-    seqRef.current += 1;
-    const name = `pasted-${seqRef.current}.${imageExtFromMime(f.type)}`;
-    return new File([f], name, {
-      type: f.type || "image/png",
-      lastModified: f.lastModified,
-    });
-  });
-}
-
-/** Local preview URL for an attached image; revoked when the file changes or the chip unmounts. */
-function useImageObjectUrl(file: File): string | null {
-  const url = useMemo(() => {
-    if (!file.type.startsWith("image/")) return null;
-    if (
-      typeof URL === "undefined" ||
-      typeof URL.createObjectURL !== "function"
-    ) {
-      return null;
-    }
-    return URL.createObjectURL(file);
-  }, [file]);
-  useEffect(() => {
-    if (!url) return;
-    return () => {
-      if (
-        typeof URL !== "undefined" &&
-        typeof URL.revokeObjectURL === "function"
-      ) {
-        URL.revokeObjectURL(url);
-      }
-    };
-  }, [url]);
-  return url;
-}
-
-/** Live attachment chip; image files render a thumbnail instead of the generic icon. */
-function AttachedFileChip({
-  file,
-  disabled,
-  onRemove,
-}: {
-  file: File;
-  disabled: boolean;
-  onRemove: () => void;
-}) {
-  const { t } = useT();
-  const { svg, label } = fileTypeIcon(file.type, file.name);
-  const thumbUrl = useImageObjectUrl(file);
-  const tip = t("composer.attachmentTooltip", {
-    fileName: file.name,
-    label,
-    size: fmtBytes(file.size, t),
-  });
-  return (
-    <span
-      className={[
-        "composer-attachment-chip",
-        thumbUrl ? "composer-attachment-chip--image" : "",
-        disabled ? "composer-attachment-chip--disabled" : "",
-      ]
-        .filter(Boolean)
-        .join(" ")}
-      title={tip}
-      aria-disabled={disabled ? "true" : undefined}
-      data-testid="composer-attachment-chip"
-    >
-      <span className="composer-attachment-chip-icon" aria-hidden="true">
-        {thumbUrl ? (
-          <img
-            className="composer-attachment-thumb"
-            src={thumbUrl}
-            alt=""
-            data-testid="composer-attachment-thumb"
-          />
-        ) : (
-          svg
-        )}
-      </span>
-      <span className="composer-attachment-chip-name">{file.name}</span>
-      <button
-        type="button"
-        className="composer-attachment-chip-remove"
-        aria-label={t("composer.removeAttachment", { fileName: file.name })}
-        onClick={onRemove}
-      >
-        ×
-      </button>
-    </span>
-  );
-}
-
-type SlashRow = { name: string; description: string };
-
-type WorkspaceFileRow = { name: string; path_rel: string; kind: string };
-
-/** Floating slash menu anchored to **`composer-field-wrap`** (viewport-relative). */
-type PickerFloatRect = {
-  left: number;
-  width: number;
-  bottom: number;
-  maxH: number;
-};
+import { AttachedFileChip } from "./AttachedFileChip";
+import {
+  clipboardImageFiles,
+  renamePastedImages,
+} from "./composerAttachments";
+import {
+  clamp01,
+  displayLlmId,
+  displayModeLabel,
+  fmtInt,
+} from "./composerLabels";
+import { ComposerModeMenus } from "./ComposerModeMenus";
+import { SlashAtPickerMenus } from "./SlashAtPickerMenus";
+import { useComposerSheetLayout } from "./useComposerSheetLayout";
+import { useEnhancePrompt } from "./useEnhancePrompt";
+import { useSlashAtPickers } from "./useSlashAtPickers";
 
 // Outline treatment per session mode (styles.css .composer-tab.mode-*); anything
 // unknown falls back to the agent look.
@@ -292,9 +113,6 @@ export function Composer(props: {
   );
   /** Screen rect of the open trigger, so the portaled menu (frosted glass over chat) can anchor to it. */
   const [menuAnchorRect, setMenuAnchorRect] = useState<DOMRect | null>(null);
-  /** Live query for the model menu filter (only meaningful while `menuOpen === "llm"`). */
-  const [llmQuery, setLlmQuery] = useState("");
-  const llmFilterRef = useRef<HTMLInputElement | null>(null);
   const [contextPopoverOpen, setContextPopoverOpen] = useState(false);
   /** After closing the breakdown, hide hover tooltip until pointer leaves the ring. */
   const [contextTipSuppressed, setContextTipSuppressed] = useState(false);
@@ -339,70 +157,43 @@ export function Composer(props: {
     [],
   );
   const [composerScrollTop, setComposerScrollTop] = useState(0);
-  /** True while the prompt-improvement request is in flight. */
-  const [enhancing, setEnhancing] = useState(false);
-  /** Request failure shown without changing the user's draft. */
-  const [enhanceErr, setEnhanceErr] = useState<string | null>(null);
-  /** Draft saved just before improving it so Ctrl+Z can restore it once. */
-  const preEnhanceRef = useRef<string | null>(null);
-  /** Bump when the slash draft changes or is dismissed so stale list responses are ignored. */
-  const slashFetchGenRef = useRef(0);
-  const [slashItems, setSlashItems] = useState<SlashRow[]>([]);
-  /** Index of the keyboard-highlighted row across the flat skills+commands list. */
-  const [slashActive, setSlashActive] = useState(0);
-  /** Built-in deterministic commands (/compact, /plugin) shown as a separate group. */
-  const [commandItems, setCommandItems] = useState<SlashRow[]>([]);
-  const commandItemsRef = useRef<SlashRow[]>([]);
-  const commandsFetchedRef = useRef(false);
-  const [slashOpen, setSlashOpen] = useState(false);
-  const [slashPrefix, setSlashPrefix] = useState("");
-  const [slashLoading, setSlashLoading] = useState(false);
-  const [slashErr, setSlashErr] = useState<string | null>(null);
-  const [slashPage, setSlashPage] = useState(1);
-  const [slashHasMore, setSlashHasMore] = useState(false);
-  const [slashReplace, setSlashReplace] = useState<{
-    from: number;
-    to: number;
-  } | null>(null);
-  const [pickerFloatRect, setPickerFloatRect] =
-    useState<PickerFloatRect | null>(null);
-  /** Server returned zero rows for failed `prefix`; hide picker/chip while the user extends that prefix at the same `/`. */
-  const [slashNoMatch, setSlashNoMatch] = useState<{
-    slashIdx: number;
-    prefix: string;
-  } | null>(null);
-  const atFetchGenRef = useRef(0);
-  /**
-   * After a workspace row is chosen, `setSelectionRange` + textarea `select` fires
-   * `updatePickerMenus` while the line still matches `atMenuDraftAtCaret`
-   * (file picks append a trailing space, which MENU_PATH treats as inside the `@` token).
-   * Skip reopening `@` on the next picker sync ticks (handles duplicate selection events).
-   */
-  const deferAtDraftPickerTicksRef = useRef(0);
-  const [atItems, setAtItems] = useState<WorkspaceFileRow[]>([]);
-  const [atOpen, setAtOpen] = useState(false);
-  const [atPrefix, setAtPrefix] = useState("");
-  const [atLoading, setAtLoading] = useState(false);
-  const [atErr, setAtErr] = useState<string | null>(null);
-  const [atPage, setAtPage] = useState(1);
-  const [atHasMore, setAtHasMore] = useState(false);
-  const [atReplace, setAtReplace] = useState<{
-    from: number;
-    to: number;
-  } | null>(null);
-  const [atNoMatch, setAtNoMatch] = useState<{
-    atIdx: number;
-    prefix: string;
-  } | null>(null);
-  const [caretPos, setCaretPos] = useState(0);
-  /** Stacked-shell viewports (`max-width`) use a bottom sheet so the picker is not clipped off-screen. */
-  const [pickerUseSheet, setPickerUseSheet] = useState(() => {
-    if (typeof window === "undefined") {
-      return false;
-    }
-    return window.matchMedia(shellStackMaxWidthMediaQuery).matches;
+  const {
+    enhancing,
+    enhanceErr,
+    clearEnhanceState,
+    enhancePrompt,
+    restorePreEnhanceDraft,
+  } = useEnhancePrompt({
+    value: props.value,
+    onChange: props.onChange,
+    sessionId: props.sessionId || "",
+    generating: props.generating === true,
+    taRef,
   });
-  const [sheetBottomPx, setSheetBottomPx] = useState<number | null>(null);
+  const picker = useSlashAtPickers({
+    sessionId: props.sessionId || "",
+    value: props.value,
+    onChange: props.onChange,
+    taRef,
+  });
+  const {
+    slashOpen,
+    atOpen,
+    pickerOpen,
+    slashItems,
+    commandMatches,
+    slashRows,
+    slashActiveIdx,
+    setSlashActive,
+    atItems,
+    slashNoMatch,
+    atNoMatch,
+    updatePickerMenus,
+    dismissSlashAtPickers,
+    applySlashChoice,
+    applyAtChoice,
+  } = picker;
+  const [caretPos, setCaretPos] = useState(0);
 
   const focusEpoch = props.focusEpoch ?? 0;
   /** Tracks session id for docked composer so switching chats in History refocuses input. */
@@ -440,77 +231,16 @@ export function Composer(props: {
     el.focus();
   }, [props.isEmpty, props.sessionId]);
 
-  const pickerOpen = slashOpen || atOpen;
   const sheetOverlayOpen = pickerOpen || contextPopoverOpen;
 
-  const measureSheetBottom = useCallback(() => {
-    if (typeof window === "undefined") {
-      return;
-    }
-    const useSheet = window.matchMedia(shellStackMaxWidthMediaQuery).matches;
-    if (!useSheet) {
-      setSheetBottomPx(null);
-      return;
-    }
-    if (props.isEmpty) {
-      setSheetBottomPx(0);
-      return;
-    }
-    const el =
-      composerCardRef.current ??
-      document.querySelector<HTMLElement>(
-        ".composer-wrap-docked .composer-card",
-      );
-    if (!el) {
-      setSheetBottomPx(null);
-      return;
-    }
-    const r = el.getBoundingClientRect();
-    setSheetBottomPx(Math.max(0, Math.round(window.innerHeight - r.top + 8)));
-  }, [props.isEmpty]);
-
-  useEffect(() => {
-    if (typeof window === "undefined") {
-      return;
-    }
-    const mq = window.matchMedia(shellStackMaxWidthMediaQuery);
-    const sync = () => setPickerUseSheet(mq.matches);
-    sync();
-    mq.addEventListener("change", sync);
-    return () => mq.removeEventListener("change", sync);
-  }, []);
-
-  useLayoutEffect(() => {
-    if (!sheetOverlayOpen) {
-      setSheetBottomPx(null);
-      return;
-    }
-    if (typeof window !== "undefined") {
-      setPickerUseSheet(
-        window.matchMedia(shellStackMaxWidthMediaQuery).matches,
-      );
-    }
-    measureSheetBottom();
-    window.addEventListener("resize", measureSheetBottom);
-    window.addEventListener("scroll", measureSheetBottom, { passive: true });
-    const card =
-      composerCardRef.current ??
-      document.querySelector<HTMLElement>(
-        ".composer-wrap-docked .composer-card",
-      );
-    const ro =
-      typeof ResizeObserver !== "undefined" && card
-        ? new ResizeObserver(() => measureSheetBottom())
-        : null;
-    if (card) {
-      ro?.observe(card);
-    }
-    return () => {
-      window.removeEventListener("resize", measureSheetBottom);
-      window.removeEventListener("scroll", measureSheetBottom);
-      ro?.disconnect();
-    };
-  }, [sheetOverlayOpen, measureSheetBottom]);
+  const { pickerUseSheet, sheetBottomPx, pickerFloatRect } =
+    useComposerSheetLayout({
+      isEmpty: props.isEmpty,
+      pickerOpen,
+      sheetOverlayOpen,
+      composerCardRef,
+      composerFieldWrapRef,
+    });
 
   const closeContextPopover = useCallback(() => {
     setContextPopoverOpen(false);
@@ -523,489 +253,6 @@ export function Composer(props: {
       closeContextPopover();
     }
   }, [pickerOpen, contextPopoverOpen, closeContextPopover]);
-  const measurePickerFloat = useCallback(() => {
-    if (!pickerOpen) {
-      setPickerFloatRect(null);
-      return;
-    }
-    if (pickerUseSheet) {
-      setPickerFloatRect(null);
-      return;
-    }
-    const el = composerFieldWrapRef.current;
-    if (!el) {
-      setPickerFloatRect(null);
-      return;
-    }
-    const r = el.getBoundingClientRect();
-    if (r.width < 8) {
-      setPickerFloatRect(null);
-      return;
-    }
-    const maxH = Math.min(260, Math.round(window.innerHeight * 0.42));
-    setPickerFloatRect({
-      left: r.left,
-      width: r.width,
-      bottom: window.innerHeight - r.top + 8,
-      maxH,
-    });
-  }, [pickerOpen, pickerUseSheet]);
-
-  useLayoutEffect(() => {
-    if (!pickerOpen) {
-      setPickerFloatRect(null);
-      return;
-    }
-    if (pickerUseSheet) {
-      setPickerFloatRect(null);
-      return;
-    }
-    measurePickerFloat();
-    const el = composerFieldWrapRef.current;
-    let ro: ResizeObserver | null = null;
-    if (typeof ResizeObserver !== "undefined" && el) {
-      ro = new ResizeObserver(() => measurePickerFloat());
-      ro.observe(el);
-    }
-    window.addEventListener("resize", measurePickerFloat);
-    const onMsgs = () => measurePickerFloat();
-    const shellMobile =
-      typeof document !== "undefined" &&
-      window.matchMedia(shellStackMaxWidthMediaQuery).matches;
-    if (shellMobile) {
-      window.addEventListener("scroll", onMsgs, { passive: true });
-    } else {
-      const msgEl =
-        typeof document !== "undefined"
-          ? document.getElementById("messages")
-          : null;
-      msgEl?.addEventListener("scroll", onMsgs, { passive: true });
-    }
-    return () => {
-      ro?.disconnect();
-      window.removeEventListener("resize", measurePickerFloat);
-      if (shellMobile) {
-        window.removeEventListener("scroll", onMsgs);
-      } else {
-        const msgEl =
-          typeof document !== "undefined"
-            ? document.getElementById("messages")
-            : null;
-        msgEl?.removeEventListener("scroll", onMsgs);
-      }
-    };
-  }, [pickerOpen, pickerUseSheet, measurePickerFloat, props.isEmpty]);
-
-  useEffect(() => {
-    if (typeof window === "undefined") {
-      return;
-    }
-    const mq = window.matchMedia(shellStackMaxWidthMediaQuery);
-    const sync = () => setPickerUseSheet(mq.matches);
-    mq.addEventListener("change", sync);
-    return () => mq.removeEventListener("change", sync);
-  }, []);
-
-  const bumpSlashFetchGen = () => {
-    slashFetchGenRef.current++;
-  };
-
-  const bumpAtFetchGen = () => {
-    atFetchGenRef.current++;
-  };
-
-  /** Close floating slash/workspace pickers without mutating textarea text (Escape or sheet backdrop). */
-  function dismissSlashAtPickers() {
-    setSlashOpen(false);
-    setSlashReplace(null);
-    setSlashNoMatch(null);
-    bumpSlashFetchGen();
-    setSlashLoading(false);
-    setSlashErr(null);
-
-    setAtOpen(false);
-    setAtReplace(null);
-    setAtNoMatch(null);
-    bumpAtFetchGen();
-    setAtLoading(false);
-    setAtErr(null);
-  }
-
-  const fetchSlashPage = useCallback(
-    async (prefix: string, page: number) => {
-      const sp = new URLSearchParams({
-        page: String(page),
-        page_size: "30",
-      });
-      if (prefix) {
-        sp.set("prefix", prefix);
-      }
-      const headers: Record<string, string> = {};
-      const sid = (props.sessionId || "").trim();
-      if (sid) {
-        headers["X-Coddy-Session-ID"] = sid;
-      }
-      const res = await fetch(`/coddy/slash-commands?${sp.toString()}`, {
-        headers,
-      });
-      if (!res.ok) {
-        throw new Error(`HTTP ${res.status}`);
-      }
-      return (await res.json()) as {
-        items: SlashRow[];
-        has_more: boolean;
-        page: number;
-      };
-    },
-    [props.sessionId],
-  );
-
-  // Built-in deterministic commands (/compact, /plugin) are static per config, so
-  // fetch them once the first time the slash menu opens and cache the result.
-  const fetchCommandsOnce = useCallback(async () => {
-    if (commandsFetchedRef.current) {
-      return;
-    }
-    commandsFetchedRef.current = true;
-    try {
-      const res = await fetch("/coddy/commands");
-      if (!res.ok) {
-        return;
-      }
-      const body = (await res.json()) as { items?: SlashRow[] };
-      const rows = body.items || [];
-      commandItemsRef.current = rows;
-      setCommandItems(rows);
-    } catch {
-      // Built-in commands are optional; ignore fetch errors.
-    }
-  }, []);
-
-  // Load the built-in commands once on mount so they are ready before the user
-  // narrows the slash prefix (avoids racing the skills-zero auto-close).
-  useEffect(() => {
-    void fetchCommandsOnce();
-  }, [fetchCommandsOnce]);
-
-  const fetchAtPage = useCallback(
-    async (prefix: string, page: number) => {
-      const sp = new URLSearchParams({
-        page: String(page),
-        page_size: "10",
-        prefix,
-        dirs: "true",
-      });
-      const headers: Record<string, string> = {};
-      const sid = (props.sessionId || "").trim();
-      if (sid) {
-        headers["X-Coddy-Session-ID"] = sid;
-      }
-      const res = await fetch(`/coddy/workspace/files?${sp.toString()}`, {
-        headers,
-      });
-      if (!res.ok) {
-        throw new Error(`HTTP ${res.status}`);
-      }
-      return (await res.json()) as {
-        items: WorkspaceFileRow[];
-        has_more: boolean;
-      };
-    },
-    [props.sessionId],
-  );
-
-  const enhancePrompt = useCallback(async () => {
-    if (enhancing || props.generating) {
-      return;
-    }
-    const draft = props.value.trim();
-    if (!draft) {
-      return;
-    }
-    preEnhanceRef.current = props.value;
-    setEnhancing(true);
-    setEnhanceErr(null);
-    try {
-      const headers: Record<string, string> = {
-        "Content-Type": "application/json",
-      };
-      const sessionID = (props.sessionId || "").trim();
-      if (sessionID) {
-        headers["X-Coddy-Session-ID"] = sessionID;
-      }
-      const response = await fetch("/coddy/enhance-prompt", {
-        method: "POST",
-        headers,
-        body: JSON.stringify({ text: draft }),
-      });
-      if (!response.ok) {
-        setEnhanceErr(
-          response.status === 503
-            ? t("composer.enhanceNoModel")
-            : t("composer.enhanceFailed"),
-        );
-        preEnhanceRef.current = null;
-        return;
-      }
-      const body = (await response.json()) as { text?: string };
-      const enhanced = (body.text || "").trim();
-      if (enhanced) {
-        props.onChange(enhanced);
-      } else {
-        preEnhanceRef.current = null;
-      }
-    } catch {
-      preEnhanceRef.current = null;
-      setEnhanceErr(t("composer.enhanceFailed"));
-    } finally {
-      setEnhancing(false);
-      requestAnimationFrame(() => taRef.current?.focus());
-    }
-  }, [
-    enhancing,
-    props.generating,
-    props.value,
-    props.sessionId,
-    props.onChange,
-    t,
-  ]);
-
-  const updateSlashMenu = useCallback(
-    (value: string, caret: number) => {
-      const draft = slashMenuDraftAtCaret(value, caret);
-      if (!draft.open) {
-        bumpSlashFetchGen();
-        setSlashOpen(false);
-        setSlashReplace(null);
-        setSlashNoMatch(null);
-        setSlashLoading(false);
-        return;
-      }
-      if (slashNoMatch && draftExtendsFailedSlashPrefix(draft, slashNoMatch)) {
-        bumpSlashFetchGen();
-        setSlashOpen(false);
-        setSlashReplace(null);
-        setSlashLoading(false);
-        return;
-      }
-      setSlashOpen(true);
-      setSlashReplace({ from: draft.slashIdx, to: draft.caret });
-      setSlashPrefix(draft.prefix);
-      slashFetchGenRef.current += 1;
-      const gen = slashFetchGenRef.current;
-      void (async () => {
-        const el = taRef.current;
-        const now = el
-          ? slashMenuDraftAtCaret(
-              el.value,
-              el.selectionStart ?? el.value.length,
-            )
-          : null;
-        if (
-          gen !== slashFetchGenRef.current ||
-          !now ||
-          !now.open ||
-          now.slashIdx !== draft.slashIdx ||
-          now.prefix !== draft.prefix
-        ) {
-          return;
-        }
-        setSlashLoading(true);
-        setSlashErr(null);
-        try {
-          const body = await fetchSlashPage(now.prefix, 1);
-          if (gen !== slashFetchGenRef.current) {
-            return;
-          }
-          const el2 = taRef.current;
-          const after = el2
-            ? slashMenuDraftAtCaret(
-                el2.value,
-                el2.selectionStart ?? el2.value.length,
-              )
-            : null;
-          if (
-            !after ||
-            !after.open ||
-            after.slashIdx !== now.slashIdx ||
-            after.prefix !== now.prefix
-          ) {
-            return;
-          }
-          const rows = body.items || [];
-          setSlashItems(rows);
-          setSlashPage(1);
-          setSlashHasMore(!!body.has_more);
-          if (rows.length === 0) {
-            // No skills match — but keep the menu open if a built-in command does.
-            const cmdMatches = filterCommandRows(
-              commandItemsRef.current,
-              after.prefix,
-            );
-            if (cmdMatches.length === 0) {
-              setSlashNoMatch({
-                slashIdx: after.slashIdx,
-                prefix: after.prefix,
-              });
-              setSlashOpen(false);
-              setSlashReplace(null);
-            } else {
-              setSlashNoMatch(null);
-            }
-          } else {
-            setSlashNoMatch(null);
-          }
-        } catch (e) {
-          if (gen !== slashFetchGenRef.current) {
-            return;
-          }
-          setSlashErr(
-            e instanceof Error ? e.message : t("composer.requestFailed"),
-          );
-          setSlashItems([]);
-          setSlashHasMore(false);
-          setSlashNoMatch(null);
-        } finally {
-          if (gen === slashFetchGenRef.current) {
-            setSlashLoading(false);
-          }
-        }
-      })();
-    },
-    [fetchSlashPage, slashNoMatch, t],
-  );
-
-  const updateAtMenu = useCallback(
-    (value: string, caret: number) => {
-      const draft = atMenuDraftAtCaret(value, caret);
-      if (!draft.open) {
-        bumpAtFetchGen();
-        setAtOpen(false);
-        setAtReplace(null);
-        setAtNoMatch(null);
-        setAtLoading(false);
-        return;
-      }
-      if (atNoMatch && draftExtendsFailedAtPrefix(draft, atNoMatch)) {
-        bumpAtFetchGen();
-        setAtOpen(false);
-        setAtReplace(null);
-        setAtLoading(false);
-        return;
-      }
-      setAtOpen(true);
-      setAtReplace({ from: draft.atIdx, to: draft.caret });
-      setAtPrefix(draft.prefix);
-
-      if (draft.prefix.trim() === "") {
-        bumpAtFetchGen();
-        const wk =
-          (props.sessionId || "").trim() || WORKSPACE_AT_RECENTS_NO_SESSION_KEY;
-        const recents = readWorkspaceAtRecents(wk).map(pickerRowFromRecent);
-        setAtItems(recents);
-        setAtPage(1);
-        setAtHasMore(false);
-        setAtNoMatch(null);
-        setAtLoading(false);
-        setAtErr(null);
-        return;
-      }
-
-      atFetchGenRef.current += 1;
-      const gen = atFetchGenRef.current;
-      void (async () => {
-        const el = taRef.current;
-        const now = el
-          ? atMenuDraftAtCaret(el.value, el.selectionStart ?? el.value.length)
-          : null;
-        if (
-          gen !== atFetchGenRef.current ||
-          !now ||
-          !now.open ||
-          now.atIdx !== draft.atIdx ||
-          now.prefix !== draft.prefix
-        ) {
-          return;
-        }
-        setAtLoading(true);
-        setAtErr(null);
-        try {
-          const body = await fetchAtPage(now.prefix.trimEnd(), 1);
-          if (gen !== atFetchGenRef.current) {
-            return;
-          }
-          const el2 = taRef.current;
-          const after = el2
-            ? atMenuDraftAtCaret(
-                el2.value,
-                el2.selectionStart ?? el2.value.length,
-              )
-            : null;
-          if (
-            !after ||
-            !after.open ||
-            after.atIdx !== now.atIdx ||
-            after.prefix !== now.prefix
-          ) {
-            return;
-          }
-          const rows = body.items || [];
-          setAtItems(rows);
-          setAtPage(1);
-          setAtHasMore(!!body.has_more);
-          if (rows.length === 0) {
-            setAtNoMatch({ atIdx: after.atIdx, prefix: after.prefix });
-            setAtItems([]);
-            setAtHasMore(false);
-          } else {
-            setAtNoMatch(null);
-          }
-        } catch (e) {
-          if (gen !== atFetchGenRef.current) {
-            return;
-          }
-          setAtErr(
-            e instanceof Error ? e.message : t("composer.requestFailed"),
-          );
-          setAtItems([]);
-          setAtHasMore(false);
-          setAtNoMatch(null);
-        } finally {
-          if (gen === atFetchGenRef.current) {
-            setAtLoading(false);
-          }
-        }
-      })();
-    },
-    [fetchAtPage, atNoMatch, props.sessionId, t],
-  );
-
-  const updatePickerMenus = useCallback(
-    (value: string, caret: number) => {
-      let deferAtDraft = false;
-      if (deferAtDraftPickerTicksRef.current > 0) {
-        deferAtDraftPickerTicksRef.current -= 1;
-        deferAtDraft = true;
-      }
-      const ad = atMenuDraftAtCaret(value, caret);
-      if (ad.open && !deferAtDraft) {
-        bumpSlashFetchGen();
-        setSlashOpen(false);
-        setSlashReplace(null);
-        setSlashNoMatch(null);
-        setSlashLoading(false);
-        updateAtMenu(value, caret);
-        return;
-      }
-      bumpAtFetchGen();
-      setAtOpen(false);
-      setAtReplace(null);
-      setAtNoMatch(null);
-      setAtLoading(false);
-      updateSlashMenu(value, caret);
-    },
-    [updateAtMenu, updateSlashMenu],
-  );
 
   const maskComposerText = props.value.length > 0;
   const composerSegments = useMemo(
@@ -1069,154 +316,9 @@ export function Composer(props: {
     setComposerScrollTop(ta.scrollTop);
   }
 
-  const applySlashChoice = (name: string) => {
-    if (!slashReplace) {
-      return;
-    }
-    const { from, to } = slashReplace;
-    const insert = `/${name} `;
-    const next = props.value.slice(0, from) + insert + props.value.slice(to);
-    props.onChange(next);
-    setSlashOpen(false);
-    setSlashReplace(null);
-    setSlashNoMatch(null);
-    bumpSlashFetchGen();
-    setSlashLoading(false);
-    setAtOpen(false);
-    setAtReplace(null);
-    setAtNoMatch(null);
-    bumpAtFetchGen();
-    requestAnimationFrame(() => {
-      const el = taRef.current;
-      if (!el) {
-        return;
-      }
-      const pos = from + insert.length;
-      el.focus();
-      el.setSelectionRange(pos, pos);
-    });
-  };
-
-  const applyAtChoice = (row: WorkspaceFileRow) => {
-    if (!atReplace) {
-      return;
-    }
-    deferAtDraftPickerTicksRef.current = 2;
-    const { from, to } = atReplace;
-    const insert =
-      row.kind === "dir"
-        ? `@${row.path_rel}`
-        : `@${row.path_rel.replace(/\/$/, "")} `;
-    const next = props.value.slice(0, from) + insert + props.value.slice(to);
-    props.onChange(next);
-    recordWorkspaceAtRecent(
-      (props.sessionId || "").trim() || WORKSPACE_AT_RECENTS_NO_SESSION_KEY,
-      row,
-    );
-    setAtOpen(false);
-    setAtReplace(null);
-    setAtNoMatch(null);
-    bumpAtFetchGen();
-    setSlashOpen(false);
-    setSlashReplace(null);
-    setSlashNoMatch(null);
-    bumpSlashFetchGen();
-    setSlashLoading(false);
-    requestAnimationFrame(() => {
-      const el = taRef.current;
-      if (!el) {
-        return;
-      }
-      const pos = from + insert.length;
-      el.focus();
-      el.setSelectionRange(pos, pos);
-    });
-  };
-
-  const loadMoreSlash = () => {
-    if (!slashOpen || slashLoading || !slashHasMore) {
-      return;
-    }
-    void (async () => {
-      setSlashLoading(true);
-      setSlashErr(null);
-      try {
-        const nextPage = slashPage + 1;
-        const body = await fetchSlashPage(slashPrefix, nextPage);
-        const more = body.items || [];
-        setSlashItems((prev) => [...prev, ...more]);
-        if (more.length > 0) {
-          setSlashNoMatch(null);
-        }
-        setSlashPage(nextPage);
-        setSlashHasMore(!!body.has_more);
-      } catch (e) {
-        setSlashErr(
-          e instanceof Error ? e.message : t("composer.requestFailed"),
-        );
-      } finally {
-        setSlashLoading(false);
-      }
-    })();
-  };
-
-  const loadMoreAt = () => {
-    if (!atOpen || atLoading || !atHasMore || atPrefix.trim() === "") {
-      return;
-    }
-    void (async () => {
-      setAtLoading(true);
-      setAtErr(null);
-      try {
-        const nextPage = atPage + 1;
-        const body = await fetchAtPage(atPrefix.trimEnd(), nextPage);
-        const more = body.items || [];
-        setAtItems((prev) => [...prev, ...more]);
-        if (more.length > 0) {
-          setAtNoMatch(null);
-        }
-        setAtPage(nextPage);
-        setAtHasMore(!!body.has_more);
-      } catch (e) {
-        setAtErr(e instanceof Error ? e.message : t("composer.requestFailed"));
-      } finally {
-        setAtLoading(false);
-      }
-    })();
-  };
-
   const llmList = props.llmModels ?? [];
   const showLlm = llmList.length > 0;
   const llmVal = (props.llmModel || "").trim();
-  // Filter input appears once the backend list is long; vendor grouping kicks
-  // in whenever more than one vendor is configured. See llmModelMenu.ts.
-  const llmShowFilter = shouldShowLlmFilter(llmList.length);
-  const llmFiltered = useMemo(
-    () => filterLlmModels(llmList, llmQuery),
-    [llmList, llmQuery],
-  );
-  const llmGrouped = shouldGroupLlmModels(llmList);
-  const llmGroups = useMemo(
-    () => groupLlmModelsByVendor(llmFiltered),
-    [llmFiltered],
-  );
-  function renderLlmItem(mid: string) {
-    return (
-      <button
-        key={mid}
-        type="button"
-        role="menuitem"
-        title={mid}
-        className={`mode-item ${mid === llmVal ? "is-selected" : ""}`}
-        onClick={() => {
-          props.onLlmModelChange?.(mid);
-          closeMenu();
-        }}
-      >
-        {displayLlmId(mid, t("composer.model"))}
-      </button>
-    );
-  }
 
   const reasoningLevels = props.llmReasoningLevels ?? [];
   const showReasoning =
@@ -1226,24 +328,7 @@ export function Composer(props: {
     ? reasoningVal.slice(0, 1).toUpperCase() + reasoningVal.slice(1)
     : t("composer.reasoning");
 
-  function displayMode(id: string): string {
-    const m = id || "agent";
-    if (m === "plan") {
-      return t("composer.modePlan");
-    }
-    if (m === "agent") {
-      return t("composer.modeAgent");
-    }
-    if (m === "ask") {
-      return t("composer.modeAsk");
-    }
-    const i = m.lastIndexOf("/");
-    if (i >= 0 && i < m.length - 1) {
-      return m.slice(i + 1);
-    }
-    return m;
-  }
-  const modeLabel = displayMode(props.mode || "agent");
+  const modeLabel = displayModeLabel(props.mode || "agent", t);
   const llmLabel = llmVal
     ? displayLlmId(llmVal, t("composer.model"))
     : t("composer.model");
@@ -1272,7 +357,6 @@ export function Composer(props: {
   function closeMenu() {
     setMenuOpen(null);
     setMenuAnchorRect(null);
-    setLlmQuery("");
   }
 
   function toggleMenu(
@@ -1282,7 +366,6 @@ export function Composer(props: {
     if (menuOpen === type) {
       closeMenu();
     } else {
-      setLlmQuery("");
       setMenuAnchorRect(trigger.getBoundingClientRect());
       setMenuOpen(type);
     }
@@ -1313,199 +396,6 @@ export function Composer(props: {
       ]
         .filter(Boolean)
         .join("\n");
-
-  const commandMatches = useMemo(
-    () => (slashOpen ? filterCommandRows(commandItems, slashPrefix) : []),
-    [slashOpen, commandItems, slashPrefix],
-  );
-  // Flat, render-ordered list of selectable rows (skills first, then commands),
-  // used for arrow-key navigation. The highlighted index is clamped to it.
-  const slashRows = useMemo(
-    () => [...slashItems, ...commandMatches],
-    [slashItems, commandMatches],
-  );
-  const slashActiveIdx = slashRows.length
-    ? Math.min(Math.max(slashActive, 0), slashRows.length - 1)
-    : 0;
-  // Reset the highlight to the first row whenever the query changes or the menu
-  // (re)opens, so "first row selected by default" always holds.
-  useEffect(() => {
-    setSlashActive(0);
-  }, [slashPrefix, slashOpen]);
-  // Hide the Skills group when only built-in commands match, so a lone command
-  // does not sit under an empty "Skills" header.
-  const showSkillsSection =
-    slashLoading ||
-    !!slashErr ||
-    slashItems.length > 0 ||
-    commandMatches.length === 0;
-
-  const slashMenuChrome = (
-    <>
-      <div className="slash-menu-surface" aria-hidden />
-      <div
-        className="slash-menu-scroll"
-        style={{ maxHeight: pickerFloatRect?.maxH }}
-      >
-        {showSkillsSection ? (
-          <>
-            <div className="slash-menu-title">{t("composer.skillsTitle")}</div>
-            {slashLoading && slashItems.length === 0 ? (
-              <div className="slash-muted">{t("composer.loading")}</div>
-            ) : null}
-            {slashErr ? <div className="slash-err">{slashErr}</div> : null}
-            {!slashLoading && slashItems.length === 0 && !slashErr ? (
-              <div className="slash-muted">
-                {t("composer.noMatchingSkills")}
-              </div>
-            ) : null}
-            <ul className="slash-rows">
-              {slashItems.map((row, idx) => (
-                <li key={row.name}>
-                  <button
-                    type="button"
-                    role="option"
-                    aria-selected={idx === slashActiveIdx}
-                    className={`slash-row-btn${idx === slashActiveIdx ? " is-active" : ""}`}
-                    data-testid={`slash-command-row-${row.name}`}
-                    onMouseEnter={() => setSlashActive(idx)}
-                    onMouseDown={(e) => {
-                      e.preventDefault();
-                      applySlashChoice(row.name);
-                    }}
-                  >
-                    <span className="slash-row-line">
-                      <span className="slash-row-name">/{row.name}</span>
-                      {row.description ? (
-                        <>
-                          {" "}
-                          <span className="slash-row-desc">
-                            {row.description}
-                          </span>
-                        </>
-                      ) : null}
-                    </span>
-                  </button>
-                </li>
-              ))}
-            </ul>
-            {slashHasMore ? (
-              <button
-                type="button"
-                className="slash-load-more"
-                disabled={slashLoading}
-                data-testid="slash-command-more"
-                onMouseDown={(e) => e.preventDefault()}
-                onClick={() => loadMoreSlash()}
-              >
-                {slashLoading ? t("composer.loading") : t("composer.more")}
-              </button>
-            ) : null}
-          </>
-        ) : null}
-        {commandMatches.length > 0 ? (
-          <>
-            <div className="slash-menu-title">
-              {t("composer.commandsTitle")}
-            </div>
-            <ul className="slash-rows">
-              {commandMatches.map((row, idx) => {
-                const gidx = slashItems.length + idx;
-                return (
-                  <li key={`cmd-${row.name}`}>
-                    <button
-                      type="button"
-                      role="option"
-                      aria-selected={gidx === slashActiveIdx}
-                      className={`slash-row-btn${gidx === slashActiveIdx ? " is-active" : ""}`}
-                      data-testid={`command-row-${row.name}`}
-                      onMouseEnter={() => setSlashActive(gidx)}
-                      onMouseDown={(e) => {
-                        e.preventDefault();
-                        applySlashChoice(row.name);
-                      }}
-                    >
-                      <span className="slash-row-line">
-                        <span className="slash-row-name">/{row.name}</span>
-                        {row.description ? (
-                          <>
-                            {" "}
-                            <span className="slash-row-desc">
-                              {row.description}
-                            </span>
-                          </>
-                        ) : null}
-                      </span>
-                    </button>
-                  </li>
-                );
-              })}
-            </ul>
-          </>
-        ) : null}
-      </div>
-    </>
-  );
-
-  const atMenuChrome = (
-    <>
-      <div className="slash-menu-surface" aria-hidden />
-      <div
-        className="slash-menu-scroll"
-        style={{ maxHeight: pickerFloatRect?.maxH }}
-      >
-        <div className="slash-menu-title">
-          {t("composer.workspaceFilesTitle")}
-        </div>
-        {atPrefix.trim() === "" && atItems.length === 0 ? (
-          <div className="slash-muted">{t("composer.typeAfterAt")}</div>
-        ) : null}
-        {atLoading && atItems.length === 0 && atPrefix.trim() !== "" ? (
-          <div className="slash-muted">{t("composer.loading")}</div>
-        ) : null}
-        {atErr ? <div className="slash-err">{atErr}</div> : null}
-        {!atLoading &&
-        atItems.length === 0 &&
-        !atErr &&
-        atPrefix.trim() !== "" ? (
-          <div className="slash-muted">{t("composer.noFiles")}</div>
-        ) : null}
-        <ul className="slash-rows">
-          {atItems.map((row) => (
-            <li key={`${row.kind}:${row.path_rel}`}>
-              <button
-                type="button"
-                role="option"
-                className="slash-row-btn"
-                data-testid={`workspace-file-row-${row.path_rel.replace(/[^a-zA-Z0-9_-]+/g, "_")}`}
-                onMouseDown={(e) => {
-                  e.preventDefault();
-                  applyAtChoice(row);
-                }}
-              >
-                <span className="slash-row-name">@{row.path_rel}</span>
-                <span className="slash-row-desc">
-                  {workspacePickRowSubtitle(row)}
-                </span>
-              </button>
-            </li>
-          ))}
-        </ul>
-        {atHasMore ? (
-          <button
-            type="button"
-            className="slash-load-more"
-            disabled={atLoading}
-            data-testid="workspace-files-more"
-            onMouseDown={(e) => e.preventDefault()}
-            onClick={() => loadMoreAt()}
-          >
-            {atLoading ? t("composer.loading") : t("composer.more")}
-          </button>
-        ) : null}
-      </div>
-    </>
-  );
 
   return (
     <>
@@ -1693,8 +583,7 @@ export function Composer(props: {
                   const v = ev.target.value;
                   const caret = ev.target.selectionStart ?? v.length;
                   setCaretPos(caret);
-                  preEnhanceRef.current = null;
-                  setEnhanceErr(null);
+                  clearEnhanceState();
                   props.onChange(v);
                   updatePickerMenus(v, caret);
                 }}
@@ -1749,14 +638,14 @@ export function Composer(props: {
                   if (
                     ev.key === "z" &&
                     (ev.metaKey || ev.ctrlKey) &&
-                    !ev.shiftKey &&
-                    preEnhanceRef.current !== null
+                    !ev.shiftKey
                   ) {
-                    ev.preventDefault();
-                    const restored = preEnhanceRef.current;
-                    preEnhanceRef.current = null;
-                    props.onChange(restored);
-                    return;
+                    const restored = restorePreEnhanceDraft();
+                    if (restored !== null) {
+                      ev.preventDefault();
+                      props.onChange(restored);
+                      return;
+                    }
                   }
                   if (ev.key === "Escape" && contextPopoverOpen) {
                     ev.preventDefault();
@@ -2090,198 +979,31 @@ export function Composer(props: {
           breakdown={props.contextBreakdown}
         />
       ) : null}
-      {menuOpen && (menuUseSheet || menuAnchorRect)
-        ? createPortal(
-            <>
-              <button
-                type="button"
-                className={`mode-menu-backdrop ${menuUseSheet ? "mode-menu-backdrop--scrim" : ""}`}
-                aria-hidden="true"
-                tabIndex={-1}
-                onMouseDown={(e) => {
-                  e.preventDefault();
-                  closeMenu();
-                }}
-              />
-              <div
-                className={`mode-menu ${menuUseSheet ? "mode-menu--sheet" : `mode-menu--portal ${modeMenuDirClass}`} ${menuOpen === "llm" ? "mode-menu--llm" : ""}`}
-                role="menu"
-                style={
-                  menuUseSheet || !menuAnchorRect
-                    ? undefined
-                    : modeMenuDirClass === "opens-up"
-                      ? {
-                          left: menuAnchorRect.left,
-                          bottom: window.innerHeight - menuAnchorRect.top + 8,
-                        }
-                      : {
-                          left: menuAnchorRect.left,
-                          top: menuAnchorRect.bottom + 8,
-                        }
-                }
-              >
-                {menuOpen === "mode"
-                  ? props.modes.map((m) => (
-                      <button
-                        key={m}
-                        type="button"
-                        role="menuitem"
-                        className={`mode-item ${m === props.mode ? "is-selected" : ""}`}
-                        onClick={() => {
-                          props.onModeChange(m);
-                          closeMenu();
-                        }}
-                      >
-                        {displayMode(m)}
-                      </button>
-                    ))
-                  : null}
-                {menuOpen === "llm" ? (
-                  <>
-                    {llmShowFilter ? (
-                      <input
-                        ref={llmFilterRef}
-                        type="text"
-                        className="mode-menu-filter"
-                        data-testid="model-menu-filter"
-                        aria-label={t("composer.filterModels")}
-                        placeholder={t("composer.filterModelsPlaceholder")}
-                        autoFocus
-                        value={llmQuery}
-                        onChange={(e) => setLlmQuery(e.target.value)}
-                        onKeyDown={(e) => {
-                          if (e.key === "Escape") {
-                            e.preventDefault();
-                            e.stopPropagation();
-                            closeMenu();
-                          } else if (e.key === "Enter") {
-                            e.preventDefault();
-                            e.stopPropagation();
-                            const first = llmFiltered[0];
-                            if (first) {
-                              props.onLlmModelChange?.(first);
-                              closeMenu();
-                            }
-                          }
-                        }}
-                      />
-                    ) : null}
-                    <div className="mode-menu-scroll">
-                      {llmFiltered.length === 0 ? (
-                        <div
-                          className="mode-menu-empty"
-                          data-testid="model-menu-empty"
-                        >
-                          {t("composer.noModelsMatch", {
-                            query: llmQuery.trim(),
-                          })}
-                        </div>
-                      ) : llmGrouped ? (
-                        llmGroups.map((g) => (
-                          <div
-                            key={g.vendor || "_"}
-                            className="mode-menu-group"
-                          >
-                            <div className="mode-menu-group-label">
-                              {g.vendor || t("composer.vendorOther")}
-                            </div>
-                            {g.models.map((mid) => renderLlmItem(mid))}
-                          </div>
-                        ))
-                      ) : (
-                        llmFiltered.map((mid) => renderLlmItem(mid))
-                      )}
-                    </div>
-                  </>
-                ) : null}
-                {menuOpen === "reasoning"
-                  ? reasoningLevels.map((lv) => (
-                      <button
-                        key={lv}
-                        type="button"
-                        role="menuitem"
-                        title={lv}
-                        className={`mode-item ${lv === reasoningVal ? "is-selected" : ""}`}
-                        onClick={() => {
-                          props.onLlmReasoningChange?.(lv);
-                          closeMenu();
-                        }}
-                      >
-                        {lv.slice(0, 1).toUpperCase() + lv.slice(1)}
-                      </button>
-                    ))
-                  : null}
-              </div>
-            </>,
-            document.body,
-          )
-        : null}
-      {pickerOpen
-        ? createPortal(
-            pickerUseSheet ? (
-              <>
-                <button
-                  type="button"
-                  className="slash-sheet-backdrop"
-                  aria-label={t("composer.closePicker")}
-                  tabIndex={-1}
-                  onMouseDown={(e) => {
-                    e.preventDefault();
-                    dismissSlashAtPickers();
-                  }}
-                />
-                <div
-                  className={[
-                    "slash-menu slash-menu--sheet",
-                    !props.isEmpty ? "slash-menu--above-composer" : "",
-                  ]
-                    .filter(Boolean)
-                    .join(" ")}
-                  data-testid={
-                    atOpen ? "workspace-files-menu" : "slash-command-menu"
-                  }
-                  role="listbox"
-                  aria-label={
-                    atOpen
-                      ? t("composer.workspaceFilesAriaLabel")
-                      : t("composer.slashCommandsAriaLabel")
-                  }
-                  style={
-                    !props.isEmpty && sheetBottomPx != null
-                      ? {
-                          bottom: sheetBottomPx,
-                          ["--context-sheet-bottom" as string]: `${sheetBottomPx}px`,
-                        }
-                      : undefined
-                  }
-                >
-                  {atOpen ? atMenuChrome : slashMenuChrome}
-                </div>
-              </>
-            ) : pickerFloatRect ? (
-              <div
-                className="slash-menu slash-menu--portal"
-                data-testid={
-                  atOpen ? "workspace-files-menu" : "slash-command-menu"
-                }
-                role="listbox"
-                aria-label={
-                  atOpen
-                    ? t("composer.workspaceFilesAriaLabel")
-                    : t("composer.slashCommandsAriaLabel")
-                }
-                style={{
-                  left: pickerFloatRect.left,
-                  width: pickerFloatRect.width,
-                  bottom: pickerFloatRect.bottom,
-                }}
-              >
-                {atOpen ? atMenuChrome : slashMenuChrome}
-              </div>
-            ) : null,
-            document.body,
-          )
-        : null}
+      {menuOpen ? (
+        <ComposerModeMenus
+          menuOpen={menuOpen}
+          menuAnchorRect={menuAnchorRect}
+          menuUseSheet={menuUseSheet}
+          modeMenuDirClass={modeMenuDirClass}
+          modes={props.modes}
+          mode={props.mode}
+          onModeChange={props.onModeChange}
+          llmModels={llmList}
+          llmModel={llmVal}
+          onLlmModelChange={props.onLlmModelChange}
+          reasoningLevels={reasoningLevels}
+          reasoning={reasoningVal}
+          onLlmReasoningChange={props.onLlmReasoningChange}
+          onClose={closeMenu}
+        />
+      ) : null}
+      <SlashAtPickerMenus
+        picker={picker}
+        pickerUseSheet={pickerUseSheet}
+        pickerFloatRect={pickerFloatRect}
+        sheetBottomPx={sheetBottomPx}
+        isEmpty={props.isEmpty}
+      />
     </>
   );
 }
